+++ packages/flutter/lib/src/material/dropdown_menu_form_field.dart
@@ -63,7 +63,7 @@ class DropdownMenuFormField<T extends Object> extends FormField<T> {
     required this.dropdownMenuEntries,
     List<TextInputFormatter>? inputFormatters,
     DropdownMenuCloseBehavior closeBehavior = DropdownMenuCloseBehavior.all,
-    int maxLines = 1,
+    double maxLines = 1.0,
     TextInputAction? textInputAction,
     super.restorationId,
     super.onSaved,
